Crate jsonrpsee_http_server

jsonrpsee-http-server

jsonrpsee-http-server is a JSON RPC HTTPS server library that’s is built for async/await.

Re-exports

pub use jsonrpsee_types as types;

pub use tracing;

Modules

response

Common builders for RPC responses. Contains common builders for hyper responses.

Structs

AccessControl

Define access on control on HTTP layer.

AccessControlBuilder

Convenience builder pattern

Host

Host type

HttpServer

An HTTP JSON RPC server.

HttpServerBuilder

Builder to create JSON-RPC HTTP server.

HttpServerHandle

Handle used to run or stop the server.

RpcModule

Sets of JSON-RPC methods can be organized into a “module“s that are in turn registered on the server or, alternatively, merged with other modules to construct a cohesive API. RpcModule wraps an additional context argument that can be used to access data during call execution.

Enums

AccessControlAllowHeaders

Headers allowed to access

AccessControlAllowOrigin

Origins allowed to access

AllowHosts

Allowed hosts for http header ‘host’

DomainsValidation

Specifies if domains should be validated.